Add --output-dir-std path option to set output path for std library tags
Sometimes, the RUST_SRC_PATH directory is read-only.  For example, I can
install the Rust source code using the rust-src package on Debian, which
will place the source code at /usr/src/rustc-*.

Therefore this patch adds a --output-dir-std option that allows the user
to set the output directory for the tags file for the standard library.